Time: 3 ms (97.05%) | Memory: 43.9 MB (57.64%) - LeetSync

+++ 437-path-sum-iii/path-sum-iii.java
@@ -17,7 +17,7 @@ class Solution {
     public int pathSum(TreeNode root, int targetSum) {
         Map<Long, Integer> map = new HashMap<>();
         map.put((long) 0, 1);
-        return pathSum(root, targetSum, (long) 0, map);
+        return pathSum(root, targetSum, 0, map);
     }
 
     public int pathSum(TreeNode root, int targetSum, long sum, Map<Long, Integer> map) {
